requirements.txt
requests==2.31.0
python-dotenv>=0.19.0
mcp>=1.6.0
canvas-sdk>=0.1.0 
